_nemo/pyproject.toml
[project]
name = "nemo-canary"
version = "0.1.0"
description = "Replace faster-whisper with Nemo (using the nvidia canary model)"
requires-python = ">=3.11"
dependencies = [
    "cuda-python>=13.0",
    "cython>=3.2.2",
    "nemo-toolkit[asr]>=2.2.1",
    "packaging>=24.1",
    "torch>=2.9.1",
]

# Specific extra url for the CUDA enabled versions of torch.
# Update it when CUDA version updates (cu130 = CUDA 13.0).
[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }